setup simulation
