+++ app/README.md
@@ -1,0 +1,3 @@
+# Ahora - PineTime Application
+
+This crate stores the application level logic for Ahora.
