+++ platform/memory.x
