+++ platform/src/battery.rs
@@ -1,3 +1,4 @@
+use ahora_app::BatteryData;
 use embassy_nrf::{
     gpio::{Input, Pull},
     peripherals::{P0_12, P0_31},
@@ -7,10 +8,6 @@ pub static BATTERY_DATA: embassy_sync::signal::Signal<
     embassy_sync::blocking_mutex::raw::ThreadModeRawMutex,
     BatteryData,
 > = embassy_sync::signal::Signal::new();
-
-pub struct BatteryData {
-    pub charging: bool,
-}
 
 #[embassy_executor::task]
 pub async fn task(charging_indication_pin: P0_12, _battery_voltage_pin: P0_31) {
