+++ platform/src/haptics.rs
