+++ sim/Cargo.toml
@@ -1,0 +1,12 @@
+[package]
+name = "ahora-sim"
+version = "0.1.0"
+edition = "2021"
+
+[dependencies]
+embedded-graphics = "0.8.1"
+embedded-graphics-simulator = "0.6.0"
+
+arrayvec = {version = "0.7.4", default-features = false }
+
+ahora-app = { path = "../app" }
