+++ sim/README.md
@@ -1,0 +1,3 @@
+# Ahora - PineTime Simulator
+
+This crate stores the Ahora / PineTime simulator.
